fix(spindle): validate CSS control characters correctly
